Break polling strategies into files
